fix: move logger interface to domain level
